extract common assertions to utils
