+++ test/demo/utils/notificationCenterUtils.ts
@@ -1,4 +1,7 @@
 import {ElementHandle} from 'puppeteer';
+import * as moment from 'moment';
+
+import {Notification} from '../../../src/client';
 
 import {OFPuppeteerBrowser} from './ofPuppeteer';
 import {fin} from './fin';
@@ -30,7 +33,7 @@ export interface NotificationCardMetadata {
 
 export interface ButtonMetadata {
     title?: string;
-    icon?: string;
+    iconUrl?: string;
 }
 
 /**
@@ -55,13 +58,13 @@ export async function getCardMetadata(sourceUuid: string, notificationId: string
     const body = await getPropertyByQueryString(card, '.body .text', 'innerHTML');
     const sourceApp = await getPropertyByQueryString(card, '.source .app-name', 'innerHTML');
     const timeString = await getPropertyByQueryString(card, '.time span', 'innerHTML');
-    const icon = await getPropertyByQueryString(card, '.source img', 'src');
+    const icon = (await getPropertyByQueryString(card, '.source img', 'src')) || '';
 
     const buttonElements = await card.$$('.button');
     const buttons = await promiseMap(buttonElements, async (button): Promise<ButtonMetadata> => {
         return {
             title: await getPropertyByQueryString(button, 'span', 'innerHTML'),
-            icon: await getPropertyByQueryString(button, 'img', 'src')
+            iconUrl: (await getPropertyByQueryString(button, 'img', 'src')) || ''
         };
     });
 
@@ -88,4 +91,20 @@ async function getPropertyByQueryString(rootElement: ElementHandle, queryString:
 
 async function promiseMap<T, R>(arr: T[], fn: (elem: T) => Promise<R>): Promise<R[]> {
     return Promise.all(arr.map(fn));
+}
+
+export async function assertDOMMatches(sourceUuid: string, note: Notification): Promise<void> {
+    const cardContent: NotificationCardMetadata | undefined = await getCardMetadata(sourceUuid, note.id);
+    expect(cardContent).not.toBeUndefined();
+
+    const expectedContent: NotificationCardMetadata = {
+        title: note.title,
+        body: note.body,
+        buttons: note.buttons,
+        icon: note.icon,
+        sourceApp: sourceUuid,
+        timeString: moment(note.date).fromNow()
+    };
+
+    expect(cardContent).toEqual(expectedContent);
 }
